+++ resource/commands/create-service.sh
@@ -2,6 +2,7 @@
 service=$(get_option '.service')
 plan=$(get_option '.plan')
 service_instance=$(get_option '.service_instance')
+broker=$(get_option '.broker')
 configuration=$(get_option '.configuration')
 tags=$(get_option '.tags')
 timeout=$(get_option '.timeout')
@@ -13,9 +14,9 @@ logger::info "Executing $(logger::highlight "$command"): $service_instance"
 cf::target "$org" "$space"
 
 if [ "true" = "$update_service" ]; then
-  cf::create_or_update_service "$service" "$plan" "$service_instance" "$configuration" "$tags"
+  cf::create_or_update_service "$service" "$plan" "$service_instance" "$broker" "$configuration" "$tags"
 else
-  cf::create_service "$service" "$plan" "$service_instance" "$configuration" "$tags"
+  cf::create_service "$service" "$plan" "$service_instance" "$broker" "$configuration" "$tags"
 fi
 
 if [ "true" = "$wait" ]; then